device-backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
boto3==1.34.162
pydantic==2.8.2
python-dotenv==1.0.1
# Cognito authentication
python-jose[cryptography]==3.3.0
requests==2.31.0
# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1
black==23.9.1
flake8==6.1.0
mypy==1.6.1